requirements.txt
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.2
django-rest-swagger==2.2.0
djangorestframework==3.9.4
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
oauthlib==3.0.1
openapi-codec==1.3.2
Pillow==6.0.0
pycodestyle==2.5.0
PySocks==1.7.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
tweepy==3.7.0
uritemplate==3.0.0
urllib3==1.25.3
